+++ Changelog.md
@@ -6,6 +6,7 @@
 ### Fixed bugs:
 
 * Fix colours for MP display
+* Fix Spirit Naga attacks - they were so broken they could crash the game
 * Ensure demon lord on level 10 matches difficulty
 
 ## Release 0.7.7 (2025-06-19)